refactor: remove cyclic dependency
